requirements.txt
flask==3.1.0
flask-login==0.6.3
flask-wtf==1.2.2
flask-session==0.8.0
psycopg2-binary==2.9.10
bcrypt==4.2.1
python-dotenv==1.0.1
openai==1.82.0
requests==2.32.3
werkzeug==3.1.3
